+++ logging.yaml
@@ -52,6 +52,9 @@ loggers:
   CENBot.logging:
     level: DEBUG
     propogate: true
+  CENBot.radio:
+    level: DEBUG
+    propogate: true
   CENBot.twitch:
     level: DEBUG
     propogate: true
